Moved index to TS
